FIX: Return only WD etcd service name, not WKS etcd
If both WD 2.2.1 and WKS 1.2.0 are installed in CP4D 3.5 (may be true
for other configurations, but this is my only experience), then the etcd
backup script finds and returns two results when determining the name
of the etcd client service.

Returned values:
```
wd-discovery-etcd-client
wks-ibm-etcd-client
```

This breaks the later input and causes the `etcd-backup-restore.sh` script
to fail/error-out and not create any kind of backup.

This commit adds an additional `grep` to return only the WD etcd client
service name by filtering for `wd-` or `discovery`