FIX: for-loop index variable conflict bug
When any backup file is larger than the configured size-to-split, the
for-loop index variable `file` is reused and causes the process to
throw an error that the file is not found (because the reference to the
outer loop variable has been overwritten by the inner loop index var).

Simple fix is to rename the second index variable. Change made in both
2.1.3 and 2.2.0 versions